Relocated some classes
